+++ 2D_ATC/Path.h
@@ -13,15 +13,15 @@ public:
 	void update();
 	void draw();
 
-	sf::Vector2f GetPointPosition(int point);
+	sf::Vector2f GetPointPosition(unsigned short point);
 	
-	void AddPoint(sf::Vector2f position, int index);
+	void AddPoint(sf::Vector2f position, unsigned short index);
 	void HidePath();
 
 	int length();
 
 private:
 	std::vector<sf::CircleShape>points;
-	std::vector<int>pointsIndex;
+	std::vector<unsigned short>pointsIndex;
 };
